[fixup] moved loading screen to it's own component
